moving components to top-level
